First scripts running up to image creation.
Modified settings to work regardless of location and user.
Successfully ran through the image creation. Did not do final
steps for push to NSIDC or archive.  Won't do these until
we switch from Mark's installation.

Also, modified .gitignore to not check in the station data files
or the pngs to GitHub repo.